Added new option to layout_export_list option.
Now it is possible to add different locations to a specific layout.

Example:
layout_export_list:
  - { url: "/your/url/here", locations: ["de","en"] }

The Exporter will append the location on the url.
In this example the Exporer will export /your/url/here/de and
/your/url/here/en.

It is necessary to add the _locale information on the
controllerfunction.

e.g:
/*
 * @Route("/your/url/here/{_locale}", name="urlhere")
 */

It is possible to set defaults within the @Route annotation.